[ESSI-1805] [ESSI-1812] skip file persistance if #encode did not output a file

+++ app/services/processors/ocr.rb
@@ -23,11 +23,14 @@ module Processors
     end
 
     # imported from hydra-derivatives
+    # modified to skip calling output_file_service if file not present
     def encode_file(file_suffix, options)
       temp_file_name = output_file(file_suffix)
       self.class.encode(source_path, options, temp_file_name)
-      output_file_service.call(File.open(temp_file_name, 'rb'), directives)
-      File.unlink(temp_file_name)
+      if File.exist?(temp_file_name)
+        output_file_service.call(File.open(temp_file_name, 'rb'), directives)
+        File.unlink(temp_file_name)
+      end
     end
 
     def options_for(_format)
